FEATURE: Add sorting ability for TaxonomyTerm children
This allows sorting of TaxonomyTerm objects, by default using a simple NumericField. If either the gridfieldextensions or sortablegridfield modules are included, then the field is hidden and the GridFields can be sorted instead.

Added both modules to the suggestions in composer.json, and updated tests to cover a simple sorting case.

Fixed travis to test against 3.1 and 3.x branches, rather than master (which is now 4.0)